+++ Kernel/Modules/AgentTicketCompose.pm
@@ -1633,6 +1633,12 @@ sub _Mask {
         }
     }
 
+    # get used calendar
+    my $Calendar = $Self->{TicketObject}->TicketCalendarGet(
+        QueueID => $Param{QueueID},
+        SLAID   => $Param{SLAID},
+    );
+
     # pending data string
     $Param{PendingDateString} = $Self->{LayoutObject}->BuildDateSelection(
         %Param,
@@ -1643,6 +1649,7 @@ sub _Mask {
         Class            => $Param{Errors}->{DateInvalid} || ' ',
         Validate         => 1,
         ValidateDateInFuture => 1,
+        Calendar             => $Calendar,
     );
 
     # Multiple-Autocomplete